requirements.txt
streamlit 
phidata
openai
google-search-results
requests